feat(Organize): 현재 상황에 맞게 필요 없는 기능 제거
** 기존 추가 계획으로 개발중이던 Front, Auth, Long-Polling 관련 기능 제거 **
