+++ src/main/java/com/sprint/mission/discodeit/service/BaseService.java
@@ -1,7 +1,5 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.service.jcf.JCFEventService;
-
 import java.util.List;
 import java.util.NoSuchElementException;
