+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFUserService.java
@@ -1,10 +1,7 @@
 package com.sprint.mission.discodeit.service.jcf;
 
-import com.sprint.mission.discodeit.auth.AuthUser;
-import com.sprint.mission.discodeit.common.Event;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
-import com.sprint.mission.discodeit.service.EventService;
 import com.sprint.mission.discodeit.service.UserService;
 
 import java.util.NoSuchElementException;
@@ -19,12 +16,10 @@ public class JCFUserService extends JCFBaseService<User, UUID, UserRepository> i
     // JCFBaseService의 repository 필드와 별도로 UserRepository 타입의 필드를 두어
     // findByUsername과 같은 고유 메서드에 쉽게 접근할 수 있도록 합니다.
     private final UserRepository userRepository;
-    private final EventService eventService;
 
-    public JCFUserService(UserRepository userRepository, JCFEventService eventService) {
+    public JCFUserService(UserRepository userRepository) {
         super(userRepository);
         this.userRepository = userRepository;
-        this.eventService = eventService;
     }
 
     @Override
@@ -69,12 +64,7 @@ public class JCFUserService extends JCFBaseService<User, UUID, UserRepository> i
         User user = findByIdNonDel(userId);
         user.goOnline();
         userRepository.save(user);
-        if (eventService != null) {
-            // User 엔티티 대신 필요한 정보만 담은 AuthUser를 이벤트 데이터로 사용
-            AuthUser authUser = AuthUser.from(user);
-            Event<AuthUser> event = new Event<>(Event.EventType.USER_STATUS_CHANGED, authUser);
-            eventService.publishEvent(event);
-        }
+
     }
 
     @Override
@@ -82,12 +72,6 @@ public class JCFUserService extends JCFBaseService<User, UUID, UserRepository> i
         User user = findByIdNonDel(userId);
         user.goOffline();
         userRepository.save(user);
-
-        if (eventService != null) {
-            AuthUser authUser = AuthUser.from(user);
-            Event<AuthUser> event = new Event<>(Event.EventType.USER_STATUS_CHANGED, authUser);
-            eventService.publishEvent(event);
-        }
     }
 
     @Override
