+++ src/main/resources/config.properties
@@ -9,6 +9,3 @@ data.path.channel_messages=channel_messages.json
 data.path.direct_messages=direct_messages.json
 
 # FXML View file paths (from resources root)
-view.login=/view/LoginView.fxml
-view.register=/view/RegisterView.fxml
-view.main=/view/MainView.fxml
