+++ src/test/java/com/sprint/mission/discodeit/JCFUserServiceAdvancedTest.java
@@ -2,7 +2,6 @@ package com.sprint.mission.discodeit;
 
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.jcf.JCFUserRepository;
-import com.sprint.mission.discodeit.service.jcf.JCFEventService;
 import com.sprint.mission.discodeit.service.jcf.JCFUserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -19,13 +18,11 @@ class JCFUserServiceAdvancedTest {
 
     private JCFUserService userService;
     private JCFUserRepository userRepository;
-    private JCFEventService eventService;
 
     @BeforeEach
     void setUp() {
         userRepository = new JCFUserRepository();
-        eventService = new JCFEventService();
-        userService = new JCFUserService(userRepository,eventService);
+        userService = new JCFUserService(userRepository);
     }
 
     @Nested
