+++ src/test/java/com/sprint/mission/discodeit/JCFUserServiceTest.java
@@ -2,7 +2,6 @@ package com.sprint.mission.discodeit;
 
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.jcf.JCFUserRepository;
-import com.sprint.mission.discodeit.service.jcf.JCFEventService;
 import com.sprint.mission.discodeit.service.jcf.JCFUserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -18,14 +17,12 @@ class JCFUserServiceTest {
 
     private JCFUserService userService;
     private JCFUserRepository userRepository;
-    private JCFEventService eventService;
 
     @BeforeEach
     void setUp() {
         // 각 테스트는 독립적으로 실행되어야 하므로, 매번 새로운 Repository와 Service 객체를 생성합니다.
         userRepository = new JCFUserRepository();
-        eventService = new JCFEventService();
-        userService = new JCFUserService(userRepository, eventService);
+        userService = new JCFUserService(userRepository);
     }
 
     @Test
